@@ -22,4 +22,4 @@ require (
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 )
 
-go 1.17
+go 1.18
