@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
@@ -1023,14 +1022,14 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 			}
 
 			if testCase.EnableWebIdentityEnvVars || testCase.EnableWebIdentityConfig {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-web-identity-token-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-web-identity-token-file")
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary web identity token file: %s", err)
 				}
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(servicemocks.MockWebIdentityToken), 0600)
+				err = os.WriteFile(file.Name(), []byte(servicemocks.MockWebIdentityToken), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing web identity token file: %s", err)
@@ -1061,7 +1060,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1069,7 +1068,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1079,7 +1078,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 			}
 
 			if testCase.SharedCredentialsFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-credentials-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-credentials-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared credentials file: %s", err)
@@ -1087,7 +1086,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedCredentialsFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedCredentialsFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared credentials file: %s", err)
@@ -1277,7 +1276,7 @@ max_attempts = 10
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1285,7 +1284,7 @@ max_attempts = 10
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1454,7 +1453,7 @@ use_fips_endpoint = true
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1462,7 +1461,7 @@ use_fips_endpoint = true
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1607,7 +1606,7 @@ func TestCustomCABundle(t *testing.T) {
 				os.Setenv(k, v)
 			}
 
-			tempdir, err := ioutil.TempDir("", "temp")
+			tempdir, err := os.MkdirTemp("", "temp")
 			if err != nil {
 				t.Fatalf("error creating temp dir: %s", err)
 			}
@@ -1640,7 +1639,7 @@ func TestCustomCABundle(t *testing.T) {
 			}
 
 			if testCase.SetSharedConfigurationFile {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1648,7 +1647,7 @@ func TestCustomCABundle(t *testing.T) {
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(
+				err = os.WriteFile(
 					file.Name(),
 					[]byte(fmt.Sprintf(`
 [default]
@@ -1664,7 +1663,7 @@ ca_bundle = %s
 			}
 
 			if testCase.SetSharedConfigurationFileToInvalid {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1672,7 +1671,7 @@ ca_bundle = %s
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(
+				err = os.WriteFile(
 					file.Name(),
 					[]byte(`
 [default]
@@ -1840,7 +1839,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 				testCase.Config.StsEndpoint = aws.StringValue(mockStsSession.Config.Endpoint)
 			}
 
-			tempdir, err := ioutil.TempDir("", "temp")
+			tempdir, err := os.MkdirTemp("", "temp")
 			if err != nil {
 				t.Fatalf("error creating temp dir: %s", err)
 			}
@@ -1848,7 +1847,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 			os.Setenv("TMPDIR", tempdir)
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1856,7 +1855,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -2103,14 +2102,14 @@ web_identity_token_file = no-such-file
 				testCase.Config.StsEndpoint = aws.StringValue(mockStsSession.Config.Endpoint)
 			}
 
-			tempdir, err := ioutil.TempDir("", "temp")
+			tempdir, err := os.MkdirTemp("", "temp")
 			if err != nil {
 				t.Fatalf("error creating temp dir: %s", err)
 			}
 			defer os.Remove(tempdir)
 			os.Setenv("TMPDIR", tempdir)
 
-			tokenFile, err := ioutil.TempFile("", "aws-sdk-go-base-web-identity-token-file")
+			tokenFile, err := os.CreateTemp("", "aws-sdk-go-base-web-identity-token-file")
 			if err != nil {
 				t.Fatalf("unexpected error creating temporary web identity token file: %s", err)
 			}
@@ -2118,7 +2117,7 @@ web_identity_token_file = no-such-file
 
 			defer os.Remove(tokenFileName)
 
-			err = ioutil.WriteFile(tokenFileName, []byte(servicemocks.MockWebIdentityToken), 0600)
+			err = os.WriteFile(tokenFileName, []byte(servicemocks.MockWebIdentityToken), 0600)
 
 			if err != nil {
 				t.Fatalf("unexpected error writing web identity token file: %s", err)
@@ -2144,7 +2143,7 @@ web_identity_token_file = no-such-file
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -2156,7 +2155,7 @@ web_identity_token_file = no-such-file
 					testCase.SharedConfigurationFile += fmt.Sprintf("web_identity_token_file = %s\n", tokenFileName)
 				}
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
